Data.jsx to many files by functions just like in back-end
