+++ musasampo-frontend/src/components/data/data.jsx
@@ -7,8 +7,6 @@ import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
 
-import './data.scss';
-
 import { withRouter} from 'react-router-dom';
 import { stringify } from 'querystring';
 import ReactPlayer from "react-player";
@@ -48,6 +46,14 @@ class Data extends React.Component
   async saveToken(data)
   {
     token2 = data;
+    
+    /*//this saves token to props
+    this.props.history.push({
+      token:data
+    });*/
+   
+   /* //this is how we get it back, works in every file
+    var token2 = this.props.location.state.token;*/
   }
 
   //User Signup in login. Gets variables for creating new user
@@ -160,113 +166,6 @@ class Data extends React.Component
 
   };
 
-
-  async createBand(nsfw, bandName, bandLogo, country)
-  {
-    async function postmethod(data)
-    {
-      var FormData = data;
-      
-      const requestOptions = {
-        method: "POST",
-        headers: { 'Content-Type': 'application/json'},
-        body: JSON.stringify(FormData)
-      }
-
-      const response = await fetch('http://localhost:9000/bands/createband/:userId',requestOptions)
-      const data2 = await response.json();
-      switch (data2)
-    {
-      case '404': alert('not found'); break;
-      case '400': alert('bad request'); break;
-      case '200': alert('done'); break;
-      default: alert('something went wrong'); 
-    }
-    }
-
-  }
-
-
-
-  async createAlbum(albumName, albumLaunchDate, albumPicture, albumGenre)
-  {
-    async function postmethod(data)
-    {
-      var FormData = data;
-      
-      const requestOptions = {
-        method: "POST",
-        headers: { 'Content-Type': 'application/json'},
-        body: JSON.stringify(FormData)
-      }
-
-      const response = await fetch('http://localhost:9000/albums/:bandId',requestOptions)
-      const data2 = await response.json();
-
-      switch (data2)
-    {
-      case '404': alert('not found'); break;
-      case '400': alert('bad request'); break;
-      case '200': alert('done'); break;
-      default: alert('something went wrong'); 
-    }
-    }
-
-    
-  }
-
-
-  async createSong(songName, MP3, MP4)
-  {
-    async function postmethod(data)
-    {
-      var FormData = data;
-      
-      const requestOptions = {
-        method: "POST",
-        headers: { 'Content-Type': 'application/json'},
-        body: JSON.stringify(FormData)
-      }
-
-      const response = await fetch('http://localhost:9000/:albumId/createsong',requestOptions)
-      const data2 = await response.json();
-
-      switch (data2)
-    {
-      case '404': alert('not found'); break;
-      case '400': alert('bad request'); break;
-      case '200': alert('done'); break;
-      default: alert('something went wrong'); 
-    }
-    }
-
-    
-  }
-
-
-  //Method for Get-method for all songs from backend.
-  async getAllSongs() 
-  {
-    const response2 = await fetch('http://localhost:9000/songs')
-    const data2 = await response2.json(); 
-    
-    //this is furioiting to find right path to inside of object:
-    //alert(data2.songs[0].songName);
-    return data2;
-  }
-
-  //Method for Get-method for all bands from backend.
-  async getAllBands()
-    {
-      //BANDS INFO FROM BACK-END
-      const response1 = await fetch('http://localhost:9000/bands/')
-      const data1 = await response1.json();
-      
-      //stringify(data2.songs)
-      //alert(data1[0].country);
-          
-      return data1;
-    }
 
   //method for Allusers from backend
   async getAllUsers()
