+++ musasampo-frontend/src/components/header/header.component.jsx
@@ -136,11 +136,6 @@ class Header extends React.Component
       }
   }
 
-  checkToken()
-  {
-   return "asd";
-  }
-
   async shouldComponentUpdate()
   {
